+++ ProyectoFinal/ProyectoFinal/ProyectoFinal/SupaBase/Tablas/DetalleOrden.cs
@@ -10,7 +10,7 @@ namespace ProyectoFinal.SupaBase.Tablas
     public class DetalleOrden : BaseModel
     {
         [PrimaryKey("id_detalle")]
-        public long Id { get; set; } // Assuming int8 in the database corresponds to long in C#
+        public int Id { get; set; } 
 
         [Column("id_orden")]
         public int Id_Orden { get; set; }
